requirements.txt
certifi==2018.11.29
chardet==3.0.4
codecov==2.0.15
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
crypto==1.4.1
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11.6
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.1.0
django-reversion==3.0.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
filelock==3.0.10
future==0.15.2
httplib2==0.9.2
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
mock==2.0.0
mysqlclient==1.3.13
Naked==0.1.31
odfpy==1.4.0
openpyxl==2.5.12
pbr==5.1.1
Pillow==5.3.0
pluggy==0.8.0
py==1.7.0
pyasn1==0.4.4
pycryptodomex==3.7.2
PyJWT==1.7.1
python-alipay-sdk==1.8.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
rsa==4.0
shellescape==3.4.1
six==1.10.0
tablib==0.12.1
toml==0.10.0
tox==3.5.3
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.1.0
xlrd==1.2.0
xlwt==1.3.0
